Crate frunk_proc_macros_impl

Frunk Proc Macros

This library holds procedural macros for frunk

Links:

Source on Github

Crates.io page

Macros

Path
Build a generic path that can be used for traversals

Derive Macros

proc_macro_hack_path
Build a generic path that can be used for traversals